refactor: Controller로 분기처리 이동 및 RequestMapping에서 url 매핑
