+++ tomcat/src/main/java/nextstep/jwp/controller/AbstractController.java
@@ -1,11 +1,19 @@
 package nextstep.jwp.controller;
 
+import org.apache.coyote.http11.constant.HttpMethod;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
 
 public class AbstractController implements Controller{
 
-    public void service(HttpRequest request, HttpResponse response) throws Exception { }
+    public void service(HttpRequest request, HttpResponse response) throws Exception {
+        if (request.getMethod() == HttpMethod.GET) {
+            doGet(request, response);
+        }
+        if (request.getMethod() == HttpMethod.POST) {
+            doPost(request, response);
+        }
+    }
 
     protected void doPost(HttpRequest request, HttpResponse response) throws Exception { /* NOOP */ }
