+++ tomcat/src/main/java/nextstep/jwp/controller/BaseController.java
@@ -1,0 +1,12 @@
+package nextstep.jwp.controller;
+
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
+
+public class BaseController extends AbstractController{
+
+    @Override
+    protected void doGet(HttpRequest request, HttpResponse response) throws Exception {
+        response.loadRawString("Hello world!");
+    }
+}
