--- tomcat/src/main/java/nextstep/jwp/controller/IndexController.java
@@ -1,4 +1,0 @@
-package nextstep.jwp.controller;
-
-public class IndexController {
-}
