+++ tomcat/src/main/java/nextstep/jwp/controller/ResourceController.java
@@ -1,0 +1,12 @@
+package nextstep.jwp.controller;
+
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
+
+public class ResourceController extends AbstractController {
+
+    @Override
+    protected void doGet(HttpRequest request, HttpResponse response) throws Exception {
+        response.loadResource(request.getUrl());
+    }
+}
